Add the correct SSID for Lenovo Legion Slim 7 16ARHA7 models.
Added the correct SSID for 16ARHA7 models to cs35l41_hda_property.c
and patch_realtek.c. The SSID has been added in the correct place
as they are SSID sorted. This will fix the sound from internal
speakers on 16ARHA7 models and should be ready to be pushed
upstream.

Signed-off-by: Christian Bendiksen <christian@bendiksen.me>